Fix a bug where the password reset email wasn't being sent
This bug was occurring because the TextFormField was a widget on its own, instead of being embedded within a Form widget with key _formKey. Due to this bug, the _formKey was null since there was no Form widget at all on the page, and so validate() was getting called on null.